+++ mobile_web/controllers/subweb/index.js
@@ -1,0 +1,10 @@
+
+import express from 'express';
+let router = express.Router();
+
+import list from './list';
+
+router.route('/subweb')
+    .get(list);
+
+module.exports = router;